requirements.txt
requests>=2.31.0
websocket-client>=1.6.1
Pillow>=10.0.0